# Smaller tree