--- a/Cargo.toml
+++ b/Cargo.toml
@@ -12,6 +12,9 @@ readme = "README.md"
 
 build = "build.rs"
 
+[profile.release]
+debug = true
+
 [[bin]]
 name = "pulldown-cmark"
 required-features = ["getopts"]