--- a/Cargo.toml
+++ b/Cargo.toml
@@ -12,14 +12,10 @@ readme = "README.md"
 
 build = "build.rs"
 
-[profile.release]
-debug = true
-
 [[bin]]
 name = "pulldown-cmark"
 required-features = ["getopts"]
 doc = false
-
 
 [[bench]]
 name = "html_rendering"