--- a/benches/lib.rs
+++ b/benches/lib.rs
@@ -15,14 +15,6 @@ mod to_html {
     }
 
     #[bench]
-    fn crdt_empty_options(b: &mut test::Bencher) {
-        let input_bytes = include_bytes!("../third_party/xi-editor/crdt.md");
-        let input = from_utf8(input_bytes).unwrap();
-
-        b.iter(|| render_html(&input, Options::empty()));
-    }
-
-    #[bench]
     fn paragraph_lots_unescapes(b: &mut test::Bencher) {
         let input = "This is by far my favourite unicode code point: &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA;
         &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA; &#xAAA;